requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
easyocr==1.7.0
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
fastapi==0.115.12
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.60.1
frozenlist==1.5.0
fsspec==2024.5.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.3
google-cloud-vision==3.10.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
groq==0.31.1
grpcio==1.71.0
grpcio-status==1.71.0
gspread==6.2.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.11
imageio==2.37.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.75
langchain-experimental==0.3.4
langchain-groq==0.3.7
langchain-mongodb==0.7.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.7
langdetect==1.0.9
langsmith==0.3.45
lark==1.2.2
lazy_loader==0.4
line-bot-sdk==3.16.3
lxml==5.3.1
Markdown==3.7
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
model2vec==0.6.0
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
nltk==3.9.1
numpy==2.2.6
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.21.0
openai==1.69.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.3
pi_heif==0.22.0
pikepdf==9.5.2
pillow==12.0.0
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.1.3
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydantic-settings==2.8.1
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.11.3
PyMuPDF==1.25.4
pypandoc==1.15
pyparsing==3.2.5
pypdf==5.4.0
pypdfium2==4.30.1
pyreadline3==3.5.4
pysbd==0.3.4
pythainlp==5.1.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
ragas==0.1.17
RapidFuzz==3.12.2
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.29.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.3
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==78.1.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
starlette==0.46.1
streamlit==1.51.0
sympy==1.14.0
tenacity==9.0.0
threadpoolctl==3.6.0
tifffile==2025.8.28
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
toml==0.10.2
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.50.3
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
ultralytics==8.3.229
ultralytics-thop==2.0.18
unstructured==0.17.2
unstructured-client==0.32.0
unstructured-inference==0.8.10
unstructured.pytesseract==0.3.15
urllib3==2.5.0
uvicorn==0.34.0
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.6.0
yarl==1.18.3
zstandard==0.23.0
